feat: show warning on invalid event type combination
